Don't copy binaries from flow image
Updates the gitpod dockerfile to stop copying binaries out of the flow
docker image. That approach broke when we updated the flow image to use
the latest ubuntu image as a base, because the built binaries depend on
a newer glibc version than is installed in the gitpod base.

So this updates the gitpod dockerfile to install things the old
fashioned way, which will hopefully be more robust.